get user input working
there are still some bugs where a shape is being moved while it is already placed.
I assume this will be fixed when I stop the current shape from moving when it is
fast placed as well as improve the user input detection.

there is also a lot of code duplication, that will decrease as the project moves on.

adding rotation will definately increase duplicated code, so hopefully an ok API will
form from that